Move latest and just_update api endpoint from rubygems controller to new activites controller
