+++ test/functional/api/v1/rubygems_controller_test.rb
@@ -393,69 +393,5 @@ class Api::V1::RubygemsControllerTest < ActionController::TestCase
       end
     end
 
-    context "On GET to latest" do
-      setup do
-        @rubygem_1 = Factory(:rubygem)
-        @version_1 = Factory(:version, :rubygem => @rubygem_1)
-        @version_2 = Factory(:version, :rubygem => @rubygem_1)
-
-        @rubygem_2 = Factory(:rubygem)
-        @version_3 = Factory(:version, :rubygem => @rubygem_2)
-
-        @rubygem_3 = Factory(:rubygem)
-        @version_4 = Factory(:version, :rubygem => @rubygem_3)
-
-        stub(Rubygem).latest(50){ [@rubygem_2, @rubygem_3] }
-      end
-
-      should "return correct JSON for latest gems" do
-        get :latest, :format => :json
-        should_return_latest_gems MultiJson.decode(@response.body)
-      end
-
-      should "return correct YAML for latest gems" do
-        get :latest, :format => :yaml
-        should_return_latest_gems YAML.load(@response.body)
-      end
-
-      should "return correct XML for latest gems" do
-        get :latest, :format => :xml
-        gems = Hash.from_xml(Nokogiri.parse(@response.body).to_xml)['rubygems']
-        should_return_latest_gems(gems)
-      end
-    end
-
-    context "On GET to just_updated" do
-      setup do
-        @rubygem_1 = Factory(:rubygem)
-        @version_1 = Factory(:version, :rubygem => @rubygem_1)
-        @version_2 = Factory(:version, :rubygem => @rubygem_1)
-
-        @rubygem_2 = Factory(:rubygem)
-        @version_3 = Factory(:version, :rubygem => @rubygem_2)
-
-        @rubygem_3 = Factory(:rubygem)
-        @version_4 = Factory(:version, :rubygem => @rubygem_3)
-
-        stub(Version).just_updated(50){ [@version_2, @version_3, @version_4] }
-      end
-
-      should "return correct JSON for just_updated gems" do
-        get :just_updated, :format => :json
-        should_return_just_updated_gems MultiJson.decode(@response.body)
-      end
-
-      should "return correct YAML for just_updated gems" do
-        get :just_updated, :format => :yaml
-        should_return_just_updated_gems YAML.load(@response.body)
-      end
-
-      should "return correct XML for just_updated gems" do
-        get :just_updated, :format => :xml
-        gems = Hash.from_xml(Nokogiri.parse(@response.body).to_xml)['rubygems']
-        should_return_just_updated_gems(gems)
-      end
-    end
-
   end
 end
